cleaned up profile button
