+++ package.json
@@ -16,7 +16,6 @@
     "@typescript-eslint/parser": "^4.17.0",
     "eslint": "^7.21.0",
     "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-prettier": "^3.3.1",
     "prettier": "^2.2.1",
     "react": "^17.0.1",
     "react-bootstrap": "^1.5.2",
@@ -53,6 +52,7 @@
     ]
   },
   "devDependencies": {
+    "@types/styled-components": "^5.1.8",
     "cors": "^2.8.5",
     "eslint-plugin-react": "^7.22.0",
     "express": "^4.17.1"
