+++ src/azure/authConfig.ts
@@ -15,4 +15,5 @@ export const loginRequest: PopupRequest = {
 // Add here the endpoints for MS Graph API services you would like to use.
 export const graphConfig = {
   graphMeEndpoint: 'https://graph.microsoft.com/v1.0/me',
+  graphMePhoto: 'https://graph.microsoft.com/v1.0/me/photo'
 };
