+++ src/azure/graph.tsx
@@ -19,19 +19,8 @@ export async function callMsGraph(accessToken: string) {
 
 export const ProfileData = (props: any) => {
   return (
-    <div id="profile-div">
-      <p>
-        <strong>Title: </strong> {props.graphData.jobTitle}
-      </p>
-      <p>
-        <strong>Mail: </strong> {props.graphData.mail}
-      </p>
-      <p>
-        <strong>Phone: </strong> {props.graphData.businessPhones[0]}
-      </p>
-      <p>
-        <strong>Location: </strong> {props.graphData.officeLocation}
-      </p>
+    <div>
+      {props.graphData.mail}
     </div>
   );
 };
