+++ src/azure/ui.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-import { Navbar, Button, DropdownButton } from 'react-bootstrap';
 import { VscAccount } from 'react-icons/vsc';
 
 import {
@@ -8,40 +6,18 @@ import {
   useMsal,
 } from '@azure/msal-react';
 import { loginRequest } from './authConfig';
-import Dropdown from 'react-bootstrap/esm/Dropdown';
-
-const SignInSignOutButton = () => {
-  const { instance } = useMsal();
-  return (
-    <>
-      <AuthenticatedTemplate>
-        <VscAccount onClick={() => instance.logout()}>
-
-        </VscAccount>
-      </AuthenticatedTemplate>
-      <UnauthenticatedTemplate>
-        <VscAccount onClick={() => instance.loginPopup(loginRequest)}>
-
-        </VscAccount>
-      </UnauthenticatedTemplate>
-    </>
-  );
-};
 
 export const PageLayout = (props: any) => {
 
   const { instance } = useMsal();
   return (
     <>
-      {/* <SignInSignOutButton /> */}
       <AuthenticatedTemplate>
         <VscAccount onClick={() => instance.logout()}>
-
         </VscAccount>
       </AuthenticatedTemplate>
       <UnauthenticatedTemplate>
         <VscAccount onClick={() => instance.loginPopup(loginRequest)}>
-
         </VscAccount>
       </UnauthenticatedTemplate>
       {props.children}
